Created separate verticle for each type
