--- src/main/java/org/folio/event/DomainEventPayloadType.java
@@ -1,5 +1,0 @@
-package org.folio.event;
-
-public enum DomainEventPayloadType {
-  CREATE, UPDATE, DELETE, DELETE_ALL
-}
