+++ src/main/java/org/folio/event/EdiExportHistoryEventType.java
@@ -5,9 +5,9 @@ import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @Getter
-public enum KafkaEventType {
+public enum EdiExportHistoryEventType {
   EXPORT_HISTORY_CREATE("edi-export-history.create"),
-  INVENTORY_ITEM_CREATE("inventory.item"),;
+  INVENTORY_ITEM_CREATE("inventory.item");
 
   private final String topicName;
 }
